+++ ttp_templates/__init__.py
@@ -1,4 +1,6 @@
 from .ttp_templates import get_template
 from .ttp_templates import parse_output
+from .ttp_templates import list_getter_platforms
+from .ttp_templates import is_getter_supported
 from .ttp_templates import list_templates
 from .ttp_templates import list_templates_refs